requirements.txt
streamlit
tensorflow-cpu==2.3.0
Pillow==7.2.0
numpy==1.18.5
firebase
# For Handling firebase and pyrebase dependency issues
sseclient==0.0.22
oauth2client==4.1.3
gcloud==0.18.3
requests_toolbelt==0.9.1
python_jwt==3.2.4
pycryptodome==3.7.3
